+++ components/animations/DogAnimation.tsx
@@ -11,6 +11,7 @@ import { useDogActionAnimation } from '../../hooks/components/animations/dog/use
 
 import { dogImageAssets } from '../../hooks/components/animations/dog/dogImageAssets';
 import { useDogRefs } from '../../hooks/components/animations/dog/useDogRefs';
+import { useDogAnimationState } from '../../hooks/components/animations/dog/useDogAnimationState';
 
 const DogAnimation = forwardRef<AnimationHandle, { 
   showVesse: boolean; 
@@ -104,17 +105,6 @@ const DogAnimation = forwardRef<AnimationHandle, {
     UnhappyOrHungryWalkinganimate
   });
 
-  useEffect(() => {
-    if (!isSitting) {
-      // currentAnimation の状態に応じてアニメーションを開始
-      if (currentAnimation === 'unhappyOrHungry') {
-        startUnhappyOrHungryWalkingAnimation();
-      } else {
-        startWalkingAnimation();
-      }
-    }
-  }, [isSitting, currentAnimation]);
-
   useDogSitAnimation({  
     isSitting,
     containerRef,
@@ -166,141 +156,41 @@ const DogAnimation = forwardRef<AnimationHandle, {
     startWalkingAnimation,
     startUnhappyOrHungryWalkingAnimation});
 
+  useDogAnimationState({isSitting,
+    currentAnimation,
+    startUnhappyOrHungryWalkingAnimation,
+    startWalkingAnimation,
+    showVesse,
+    feedButtonClick,
+    showNotes,
+    yellowNoteRef,
+    blueNoteRef,
+    showBall,
+    playButtonClick,
+    showHearts,
+    heartRef,
+    heartRef2,
+    petDetails,
+    setCurrentAnimation,
+    containerRef,
+    legBackLeftRef, 
+    legBackRightRef, 
+    legFrontLeftRef, 
+    legFrontRightRef, 
+    tailRef, 
+    headFaceRef, 
+    headEyeRef, 
+    bodyRef, 
+    earRef, 
+    earRightRef, 
+    jawRef
+  });
+
   const handleClick = () => {
     if (!isSitting) {
       setIsSitting(true);
     }
   };
-
-  useEffect(() => {
-    if (showVesse) {
-      feedButtonClick();  
-    }
-  }, [showVesse]);
-
-  useEffect(() => {
-    if (showNotes && yellowNoteRef.current && blueNoteRef.current) {
-      console.log("noteTl started");
-  
-      const noteTl = gsap.timeline();
-  
-      // 音符のアニメーション
-      noteTl.to(yellowNoteRef.current, {
-        rotation: 30,
-        transformOrigin: "center",
-        duration: 1.0,
-        ease: "power1.out",
-        repeat: -1,
-        yoyo: true,
-      });
-  
-      noteTl.to(
-        blueNoteRef.current,
-        {
-          rotation: -30,
-          transformOrigin: "center",
-          duration: 1.0,
-          ease: "power1.out",
-          repeat: -1,
-          yoyo: true,
-        },
-        "<" // 同時に実行
-      );
-  
-      return () => {
-        noteTl.kill(); // クリーンアップ
-      };
-    }
-  }, [showNotes]);
-
-  useEffect(() => {
-    if (showBall) {
-      playButtonClick(); 
-    }
-  }, [showBall]);
-
-  useEffect(() => {
-    if (showHearts && heartRef.current && heartRef2.current) {
-      console.log("heartTl started");
-  
-      const heartTl = gsap.timeline();
-  
-      heartTl.to(
-        heartRef.current,
-        {
-          rotation: 30,
-          transformOrigin: 'center',
-          duration: 1.0,
-          ease: 'power1.out',
-          repeat: -1,
-          yoyo: true,
-        }
-      );
-  
-      heartTl.to(
-        heartRef2.current,
-        {
-          rotation: -30,
-          transformOrigin: 'center',
-          duration: 1.0,
-          ease: 'power1.out',
-          repeat: -1,
-          yoyo: true,
-        },
-        "<" // 同時に実行する
-      );
-  
-      return () => {
-        heartTl.kill(); // クリーンアップ
-      };
-    }
-  }, [showHearts]);
-
-  useEffect(() => {
-    console.log("petDetails:", petDetails);
-    console.log("petDetails.states:", petDetails.states);
-
-    if (petDetails) {
-      const newAnimationState = (petDetails.states & 1 || petDetails.states & 2) ? 'unhappyOrHungry' : 'normal';
-      if (currentAnimation !== newAnimationState) {
-        setCurrentAnimation(newAnimationState);
-      }
-    }
-  }, [petDetails?.states]); // `petDetails.states` の変化のみを監視
-
-  useEffect(() => {
-    // 現在のアニメーションを停止
-    gsap.killTweensOf(containerRef.current);
-    gsap.killTweensOf([
-      legBackLeftRef.current, 
-      legBackRightRef.current, 
-      legFrontLeftRef.current, 
-      legFrontRightRef.current, 
-      tailRef.current, 
-      headFaceRef.current, 
-      headEyeRef.current, 
-      bodyRef.current, 
-      earRef.current, 
-      earRightRef.current, 
-      jawRef.current
-    ]);
-
-    console.log('currentAnimation:', currentAnimation);
-  
-    // アニメーションの状態に応じて新しいアニメーションを開始
-    if (currentAnimation === 'unhappyOrHungry') {
-      console.log('unhappyOrHungry state is true');
-      startUnhappyOrHungryWalkingAnimation();
-    } else {
-      console.log('normal state is true');
-      startWalkingAnimation();
-    }
-  
-    // クリーンアップ
-    return () => {
-      gsap.killTweensOf(containerRef.current);
-    };
-  }, [currentAnimation]);
 
   return (
     <div className="dog-container relative w-[450px] h-[350px] mx-auto cursor-pointer z-50" ref={containerRef} onClick={handleClick}>
